Extract ModernMinecraftVersionProfile to separate class.
