+++ installer/src/main/kotlin/gg/essential/installer/modloader/ForgeModloader.kt
@@ -77,39 +77,6 @@ object ForgeModloader : Modloader(ModloaderType.FORGE) {
         val libraries: List<ModernMinecraftVersionProfile.Library>? = null,
     )
 
-    /**
-     * Represents the json format of the Minecraft launcher's modern version profiles
-     *
-     * It does not include the full spec, merely what is (or is thought to be) required by the installer to function
-     * Due to the missing fields, these should never be serialized to a file directly from these object.
-     * You should always use the full/raw version profile json provided by the modloader.
-     */
-    @Serializable
-    data class ModernMinecraftVersionProfile(
-        val id: MinecraftVersionProfileId,
-        val inheritsFrom: String,
-        val releaseTime: String,
-        val time: String,
-        val type: String,
-        val mainClass: String,
-        val arguments: Arguments,
-        val libraries: List<Library>,
-    ) {
-
-        @Serializable
-        data class Arguments(val game: List<String>, val jvm: List<String>)
-
-        @Serializable
-        data class Library(val name: String, val downloads: Downloads)
-
-        @Serializable
-        data class Downloads(val artifact: Artifact)
-
-        @Serializable
-        data class Artifact(val path: String? = null, val url: String?, val sha1: String, val size: Long)
-
-    }
-
     override fun getMinecraftVersionProfileId(mcVersion: MCVersion, modloaderVersion: ModloaderVersion): MinecraftVersionProfileId {
         // Old forge has been very inconsistent with names for versions, this format is how modern versions format it, with "-wrapper" at the end to not mess with normal forge
         return "$mcVersion-forge-${modloaderVersion.numeric}-wrapper".replace(Regex("-{2,}"), "-")
